services/pet-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pymongo==4.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.4.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
pydantic-settings==2.1.0
alembic==1.13.1
